Code refactor and general improvements
- Load all tasks when first logging in instead of later
- Improve the task update job
- Update Gradle to 4.9
- Update dependencies

+++ t-tasks/src/main/java/com/teo/ttasks/ui/fragments/tasks/TasksFragment.kt
@@ -319,8 +319,8 @@ class TasksFragment : DaggerFragment(), TasksView, SwipeRefreshLayout.OnRefreshL
         tasksBinding.swipeRefreshLayout.isRefreshing = false
     }
 
-    override fun onSyncDone(taskSyncCount: Int) {
-        if (taskSyncCount != 0) {
+    override fun onSyncDone(taskSyncCount: Long) {
+        if (taskSyncCount != 0L) {
             Toast.makeText(context, "Synchronized $taskSyncCount tasks", Toast.LENGTH_SHORT).show()
         }
         tasksPresenter.refreshTasks(taskListId)
